gzip middleware that does not depend on the broken one in paste
I reported a bug on the one in paste. This one is conditional on
size of the resource, it starts compressing after the result is
min_length (default 1024). Not sure if it makes sense to compress
whatever fits in a packet, except for metered media or media with
different MSS and/or low bandwidth.