Ensure that only HTTP/1.1 responses are chunked
Unless the Paste server is started as "protocol_version='HTTP/1.1'"
it will send HTTP/1.0 responses, where chunked is not acceptable.
Unfortunately there is no way to know if the server is 1.1 in WSGI.

Also, check for HTTP/1.0 clients. SERVER_PROTOCOL is really request
protocol.

Unless we send some sort of fake Content-Length, the Paste server will
close the connection each time. This has been reported to paste as
http://trac.pythonpaste.org/pythonpaste/ticket/418